Prevent the use of unknown bucket properties.
When creating a new bucket property that results in new messages being
sent to vnodes in the cluster, the bucket property needs to be known by
all nodes in the cluster before it's use.  This is specifically the case
with the `write_once` bucket property; this property results in a
message arriving at a vnode that might not know about it resulting in a
crash.

To resolve this, we need to use capabilities to ensure that a bucket
property is known by all nodes in a cluster before it's use: however,
this in itself is problematic.  How do you enforce that something be
required on nodes that do not know about it yet?

This requires a two step process:

1.) We introduce a new capability called `required_bprops`: these are
required bucket properties that must be known globally before their use.
In the `write_once` case, this property will not be avilable until all
nodes support it via the capability system.
2.) We ensure that process that's responsible for the creation of new
bucket-types always is the *most recent version of Riak* in the cluster:
this ensures that the claimant *knows* about all possible values of
things that could exist in this capability: older version of the
claimant might not know they need to enforce a restriction on a given
bucket property because it's not known about.

These two properties are essential because there is not validation
process currently on bucket properties: the only validation we have is a
call to `list_to_existing_atom` triggered as part of the deserialization
of the JSON payload from the command line when creating properties.